Complete lesson 1.2 assignment
